Add Databaseseeder Articulos

+++ database/migrations/2022_09_22_145139_create_articulos_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('codigo',10)->nullable();
             $table->string('descripcion',50)->nullable();
             $table->integer('cantidad')->nullable();
-            $table->decimal('precio', 8, 2)->nullable();
+            $table->integer('precio')->nullable();
             $table->timestamps();
         });
     }
